+++ lib/way_of_working.rb
@@ -22,6 +22,7 @@ loader.ignore(File.expand_path('way_of_working/decision_record/madr/github_audit
 loader.ignore(File.expand_path('way_of_working/inclusive_language/alex/github_audit_rule.rb', __dir__))
 loader.ignore(File.expand_path('way_of_working/pull_request_template/hdi/github_audit_rule.rb', __dir__))
 loader.ignore(File.expand_path('way_of_working/readme_badge/github_audit_rule.rb', __dir__))
+loader.ignore(File.expand_path('way_of_working/versioning/semver/github_audit_rule.rb', __dir__))
 loader.setup
 
 # This is the namespace for everything associated with the Way of Working